Extract NestedPartitionsJournalDao #45
* Extract specialized NestedPartitionsJournalDao class  from ByteArrayJournalDao
* Get rid of the `pg-journal.partitioned` config flag
