+++ core/src/main/resources/reference.conf
@@ -109,7 +109,7 @@ pg-journal {
 
   tables {
     journal {
-      # Used when pg-journal.partitioned is set to true
+      # Used by akka.persistence.postgres.journal.dao.NestedPartitionsJournalDao
       partitions {
         # The size of a single partition
         size: 10000000
@@ -140,8 +140,6 @@ pg-journal {
   }
 
   dao = "akka.persistence.postgres.journal.dao.ByteArrayJournalDao"
-
-  partitioned = false
 
   tags {
     cacheTtl = 1 hour
