+++ core/src/test/resources/general.conf
@@ -55,8 +55,6 @@ pg-journal {
     "akka.persistence.postgres.query.EventAdapterTest$TaggedAsyncEvent" = test-write-event-adapter
     "akka.persistence.postgres.query.EventAdapterTest$EventAdapted" = test-read-event-adapter
   }
-
-  partitioned = false
 }
