+++ core/src/test/resources/partitioned-application-with-hard-delete.conf
@@ -1,3 +1,3 @@
 include "plain-application-with-hard-delete.conf"
 
-pg-journal.partitioned=true
+pg-journal.dao = "akka.persistence.postgres.journal.dao.NestedPartitionsJournalDao"
