+++ core/src/test/resources/partitioned-application.conf
@@ -1,3 +1,3 @@
 include "plain-application.conf"
 
-pg-journal.partitioned=true
+pg-journal.dao = "akka.persistence.postgres.journal.dao.NestedPartitionsJournalDao"
