+++ core/src/test/resources/partitioned-shared-db-application.conf
@@ -1,3 +1,3 @@
 include "plain-shared-db-application.conf"
 
-pg-journal.partitioned=true
+pg-journal.dao = "akka.persistence.postgres.journal.dao.NestedPartitionsJournalDao"
